apps/ai-agent/requirements.txt
fastapi
uvicorn
openai==0.28.1
python-dotenv
httpx
pydantic         
aiofiles  
PyJWT
starlette>=0.27.0